20241226 option btn and confirm modal

+++ src/pages/puzzle/operation-btns.tsx
@@ -5,29 +5,49 @@ import { TbReload } from "react-icons/tb";
 import { OperationBtn } from "@components/operation-btn";
 import { usePuzzle } from "@hooks/usePuzzle";
 import { useCurrentGrid } from "@hooks/useCurrentGrid";
-import { useCheckError } from "@hooks/useCheckError";
+import { useNavigate } from "react-router-dom";
 
 export const OperationBtns = () => {
-	const { deleteNumber, deleteAllNumber } = usePuzzle();
-	const { checkError } = useCheckError();
+	const navigate = useNavigate();
+	const { deleteNumber, deleteAllNumber, numberObj } = usePuzzle();
 	const { currentGrid } = useCurrentGrid();
+
+	const handleReturnToHomePage = () => {
+		let confirmed = window.confirm(
+			"Are you sure you want to return to the home page?\n(Your current progress will be lost)"
+		);
+		if (confirmed) {
+			navigate("/");
+		}
+	};
 
 	const handleDeleteNumber = () => {
 		if (!currentGrid) return;
 		deleteNumber(currentGrid);
-		checkError();
 	};
 
 	const handleDeleteAllNumber = () => {
-		if (!currentGrid) return;
-		deleteAllNumber(currentGrid);
-		checkError();
+		if (!currentGrid || !numberObj) return;
+		const gridNum = numberObj[currentGrid];
+
+		if (gridNum.num === "-") return;
+
+		let confirmed = window.confirm(
+			`Are you sure you want to remove all entered numbers ${gridNum.num}?`
+		);
+		if (confirmed) {
+			deleteAllNumber(currentGrid);
+		}
 	};
 
 	return (
 		<>
 			<div className="mb-8">
-				<OperationBtn isHighlight icon={TbReload} onClick={() => {}} />
+				<OperationBtn
+					isHighlight
+					icon={TbReload}
+					onClick={handleReturnToHomePage}
+				/>
 			</div>
 			{/* TODO : tooltips */}
 			<OperationBtn icon={HiOutlineVariable} onClick={handleDeleteAllNumber} />
